requirements.txt
google-play-scraper
python-dotenv
google-generativeai
fastapi
uvicorn
pydantic
